requirements.txt
fastapi[standard]
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
pydantic-settings
python-dotenv
openai
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
langchain-chroma
chromadb
beautifulsoup4
requests
httpx
pypdf
python-magic
aiofiles
